types: Move PlaybackState to mopidy.types
To avoid import cycles, we need this basic type to be available without
importing the full audio system.
