+++ src/mopidy/audio/__init__.py
@@ -1,12 +1,13 @@
+from mopidy.types import PlaybackState
+
 from .actor import Audio, AudioProxy
-from .constants import PlaybackState
 from .listener import AudioListener
 from .utils import supported_uri_schemes
 
 __all__ = [
     "Audio",
     "AudioProxy",
-    "PlaybackState",
+    "PlaybackState",  # Re-exported for backwards compatibility.
     "AudioListener",
     "supported_uri_schemes",
 ]
