+++ src/mopidy/audio/actor.py
@@ -12,11 +12,10 @@ from pykka.typing import ActorMemberMixin, proxy_field, proxy_method
 from mopidy import exceptions
 from mopidy.audio import tags as tags_lib
 from mopidy.audio import utils
-from mopidy.audio.constants import PlaybackState
 from mopidy.audio.listener import AudioListener
 from mopidy.internal import process
 from mopidy.internal.gi import GLib, Gst, GstPbutils
-from mopidy.types import DurationMs, Percentage
+from mopidy.types import DurationMs, Percentage, PlaybackState
 
 if TYPE_CHECKING:
     from mopidy.config import Config
