--- src/mopidy/audio/constants.py
@@ -1,14 +1,0 @@
-from enum import Enum
-
-
-class PlaybackState(str, Enum):
-    """Enum of playback states."""
-
-    #: Constant representing the paused state.
-    PAUSED = "paused"
-
-    #: Constant representing the playing state.
-    PLAYING = "playing"
-
-    #: Constant representing the stopped state.
-    STOPPED = "stopped"
